Add files via upload
Great course for anyone looking for a hands-on program that covers a wide scope of the Data Analysis process. 
In this, I have learned Data cleaning, Data analysis, Data-driven decision making, Calculation and analysis in SQL and spreadsheets, Data Visualization(Tableau), R programming, and Effective presentations.

Thanks to Google, Grow with Google, and Coursera for creating and offering such a wonderful course.